# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ClearDataResponse": {
        "description": "Report owner-scoped data deletion counts.",
        "properties": {
          "deleted_tasks": {
            "title": "Deleted Tasks",
            "type": "integer"
          }
        },
        "required": [
          "deleted_tasks"
        ],
        "title": "ClearDataResponse",
        "type": "object"
      },
      "ErrorBody": {
        "description": "Describe a machine-readable API error.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Wrap API errors in the public response envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "InstallationResponse": {
        "description": "Return anonymous installation credentials to the extension.",
        "example": {
          "created_at": "2026-09-12T18:45:00Z",
          "installation_id": "7de98bb2-6523-4f1d-b09d-caa8c92dd8f0",
          "installation_token": "inst_0123456789abcdef",
          "owner_id": "18800e89-4b37-4355-9f65-dd48f2f2bb5c"
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "installation_id": {
            "title": "Installation Id",
            "type": "string"
          },
          "installation_token": {
            "title": "Installation Token",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          }
        },
        "required": [
          "installation_id",
          "installation_token",
          "owner_id",
          "created_at"
        ],
        "title": "InstallationResponse",
        "type": "object"
      },
      "ManualTaskCreateRequest": {
        "description": "Create a manual task with an owner-scoped idempotency key.",
        "example": {
          "client_request_id": "sidebar-manual-001",
          "title": "Review Yokohama tabs before standup"
        },
        "properties": {
          "client_request_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Client Request Id",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "client_request_id",
          "title"
        ],
        "title": "ManualTaskCreateRequest",
        "type": "object"
      },
      "ManualTaskPatchRequest": {
        "description": "Patch mutable fields on a manual task.",
        "example": {
          "status": "action_complete",
          "status_reason": "Marked complete by the user.",
          "title": "Review Yokohama tabs after lunch"
        },
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "status_reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ManualTaskPatchRequest",
        "type": "object"
      },
      "ProcessingState": {
        "description": "Enumerate asynchronous processing states for a task.",
        "enum": [
          "queued",
          "running",
          "ready",
          "failed"
        ],
        "title": "ProcessingState",
        "type": "string"
      },
      "TaskListResponse": {
        "description": "Return all task cards visible to the authenticated owner.",
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "tasks"
        ],
        "title": "TaskListResponse",
        "type": "object"
      },
      "TaskOrigin": {
        "description": "Enumerate the source that created a task.",
        "enum": [
          "manual",
          "detected"
        ],
        "title": "TaskOrigin",
        "type": "string"
      },
      "TaskResponse": {
        "description": "Return a sidebar task card.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "origin": {
            "$ref": "#/components/schemas/TaskOrigin"
          },
          "processing_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error Code"
          },
          "processing_state": {
            "$ref": "#/components/schemas/ProcessingState"
          },
          "source_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Key"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "origin",
          "source_key",
          "source_url",
          "type",
          "title",
          "status",
          "status_reason",
          "summary",
          "processing_state",
          "processing_error_code",
          "observed_at",
          "updated_at"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
      "TaskStatus": {
        "description": "Enumerate user-visible task workflow states.",
        "enum": [
          "in_progress",
          "needs_attention",
          "action_complete",
          "error"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskType": {
        "description": "Enumerate supported task categories.",
        "enum": [
          "manual",
          "github",
          "research",
          "email",
          "travel",
          "shopping",
          "other"
        ],
        "title": "TaskType",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Owner-scoped API for browser task sidebar cards and onboarding.",
    "title": "Browser Task Sidebar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/data": {
      "delete": {
        "description": "Delete retained task data for the authenticated owner.\n\nArgs:\n    owner: The authenticated owner boundary.\n    repository: The repository that stores task data.\n\nReturns:\n    Counts for the deleted owner-scoped records.",
        "operationId": "clear_data_v1_data_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearDataResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear Data",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/installations": {
      "post": {
        "description": "Create an anonymous installation credential.\n\nArgs:\n    repository: The repository that persists the owner boundary.\n\nReturns:\n    The installation credential and owner identifier.",
        "operationId": "create_installation_v1_installations_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create Installation",
        "tags": [
          "installations"
        ]
      }
    },
    "/v1/tasks": {
      "get": {
        "description": "List task cards for the authenticated owner.\n\nArgs:\n    owner: The authenticated owner boundary.\n    repository: The repository that stores task data.\n\nReturns:\n    The owner's task cards.",
        "operationId": "list_tasks_v1_tasks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "description": "Create an idempotent manual task for the authenticated owner.\n\nArgs:\n    request: The manual task creation payload.\n    owner: The authenticated owner boundary.\n    repository: The repository that stores task data.\n\nReturns:\n    The created or replayed manual task.",
        "operationId": "create_manual_task_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualTaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Manual Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}": {
      "delete": {
        "description": "Delete one task owned by the authenticated owner.\n\nArgs:\n    task_id: The task identifier from the URL.\n    owner: The authenticated owner boundary.\n    repository: The repository that stores task data.\n\nReturns:\n    An empty 204 response when the task is deleted.\n\nRaises:\n    HTTPException: Raised with 404 when the task is absent for this owner.",
        "operationId": "delete_task_v1_tasks__task_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Task",
        "tags": [
          "tasks"
        ]
      },
      "patch": {
        "description": "Patch a manual task owned by the authenticated owner.\n\nArgs:\n    task_id: The task identifier from the URL.\n    request: The mutable manual task fields to update.\n    owner: The authenticated owner boundary.\n    repository: The repository that stores task data.\n\nReturns:\n    The updated task.\n\nRaises:\n    HTTPException: Raised with 400 for empty patches or 404 for absent tasks.",
        "operationId": "update_manual_task_v1_tasks__task_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualTaskPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Manual Task",
        "tags": [
          "tasks"
        ]
      }
    }
  }
}